requirements.txt
# HTTP 请求库
aiohttp>=3.9.0

# 环境变量管理
python-dotenv>=1.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据可视化
matplotlib>=3.8.0
seaborn>=0.13.0

# 图像处理
Pillow>=10.0.0

# 时区支持
pytz>=2023.3
